backend/requirements.txt
# Faredown Backend Dependencies
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
bcrypt==4.1.2

# Environment & Configuration
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# HTTP & API
httpx==0.25.2
requests==2.31.0
aiofiles==23.2.1

# AI & Machine Learning
openai==1.3.7
numpy==1.24.3
scikit-learn==1.3.2

# Data Processing
pandas==2.1.4
python-dateutil==2.8.2

# Email
emails==0.6
jinja2==3.1.2

# Caching & Sessions
redis==5.0.1
aioredis==2.0.1

# File Processing
pillow==10.1.0
python-magic==0.4.27

# Monitoring & Logging
structlog==23.2.0
sentry-sdk[fastapi]==1.38.0

# Testing (Optional)
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Development Tools
black==23.11.0
isort==5.12.0
flake8==6.1.0

# Production
gunicorn==21.2.0
